various changes i can't quite really tell about
