+++ src/lt/saltyjuice/dragas/powerbot/actions/interacting/SpinnerInteractingAction.java
@@ -1,6 +1,7 @@
 package lt.saltyjuice.dragas.powerbot.actions.interacting;
 
 import lt.saltyjuice.dragas.powerbot.Constant;
+import lt.saltyjuice.dragas.powerbot.Utility;
 import org.powerbot.script.Condition;
 import org.powerbot.script.rt4.ClientContext;
 import org.powerbot.script.rt4.GameObject;
@@ -16,12 +17,13 @@ public class SpinnerInteractingAction extends InteractingAction
     @Override
     protected int getSearchRadius()
     {
-        return 3;
+        return 12;
     }
 
     @Override
     protected void interact(GameObject obj)
     {
+        Utility.turnTo(obj);
         obj.interact("Spin");
         int delay = (int) (Math.random() * 600) + 300; //from 300 to 900 ms
         Condition.sleep(delay);
