+++ src/lt/saltyjuice/dragas/powerbot/actions/widgeting/FlaxSpinningAction.java
@@ -24,9 +24,6 @@ public class FlaxSpinningAction extends SpinInteractingAction
     protected void interact(Widget w, Component c)
     {
         super.interact(w, c);
-        long now = System.currentTimeMillis();
-        Condition.wait(() -> w.ctx.widgets.select().id(Constant.Widget.LEVEL_UP).poll().valid() ||
-        now + 20000 < System.currentTimeMillis() || isFinished(w.ctx), 1000, 21);
-        //Condition.sleep(5000);
+        Condition.wait(() -> w.ctx.widgets.select().id(Constant.Widget.LEVEL_UP).poll().valid() || isFinished(w.ctx), 1000, 64);
     }
 }
